Change MaceEvolve WinForms to fully utilise MaceEvolve.Core
