+++ MaceEvolve/Models/GraphicalCreature.cs
@@ -6,20 +6,6 @@ namespace MaceEvolve.Models
 {
     public class GraphicalCreature : Creature, IGraphical
     {
-        #region Fields
-        private Color _color = Color.Black;
-        #endregion
-
-        public Color Color
-        {
-            get
-            {
-                return _color;
-            }
-            set
-            {
-                _color = value;
-            }
-        }
+        public Color Color { get; set; }
     }
 }
